Delete circle.yml
Circle CI has become a hassle to use. It doesn't use the configuration that travis Ci uses, nor does it "auto detect" grunt and run my grunt scripts. It may be usefull to others, but I plan to remove Circle CI for now, then perhaps I'll add it later when I have the patience to figure it out.